+++ client/app/profile/profile.component.js
@@ -1,0 +1,51 @@
+'use strict';
+
+import angular from 'angular';
+import ngResource from 'angular-resource';
+import uiRouter from 'angular-ui-router';
+import routes from './profile.routes';
+import AuthModule from '../../components/auth/auth.module';
+import oauthButtons from '../../components/oauth-buttons/oauth-buttons.directive';
+
+export class ProfileController {
+  user = {
+    oldPassword: '',
+    newPassword: '',
+    confirmPassword: ''
+  };
+  errors = {
+    other: undefined
+  };
+  message = '';
+  submitted = false;
+
+
+  /*@ngInject*/
+  constructor(Auth) {
+    this.Auth = Auth;
+  }
+
+  changePassword(form) {
+    this.submitted = true;
+
+    if(form.$valid) {
+      this.Auth.changePassword(this.user.oldPassword, this.user.newPassword)
+        .then(() => {
+          this.message = 'Password successfully changed.';
+        })
+        .catch(() => {
+          form.password.$setValidity('sequelize', false);
+          this.errors.other = 'Incorrect password';
+          this.message = '';
+        });
+    }
+  }
+}
+
+export default angular.module('shyApp.profile', [ngResource, uiRouter, AuthModule, oauthButtons])
+  .config(routes)
+  .component('profile', {
+    template: require('./profile.pug'),
+    controller: ProfileController
+  })
+  .name;