+++ client/app/profile/profile.routes.js
@@ -5,6 +5,7 @@ export default function($stateProvider) {
   $stateProvider
     .state('profile', {
       url: '/profile',
-      template: '<profile></profile>'
+      template: '<profile></profile>',
+      authenticate: true
     });
 }